python/src/auth/requirements.txt
astroid==2.15.4
blinker==1.6.2
click==8.1.3
dill==0.3.6
Flask==2.3.2
Flask-MySQLdb==1.0.1
importlib-metadata==6.6.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mysqlclient==2.1.1
parso==0.8.3
platformdirs==3.5.0
PyJWT==2.6.0
pylint==2.17.4
PyMySQL==1.0.3
tomli==2.0.1
tomlkit==0.11.8
typing-extensions==4.5.0
Werkzeug==2.3.3
wrapt==1.15.0
zipp==3.15.0
